refactor/ffi: move safe_auth ffi to different crate
