+++ safe_app/build.rs
@@ -25,7 +25,8 @@ mod bindings {
     use threshold_crypto::PK_SIZE;
     use unwrap::unwrap;
 
-    const BSD_MIT_LICENSE: &str = "// Copyright 2019 MaidSafe.net limited.\n\
+    const BSD_MIT_LICENSE: &str =
+        "// Copyright 2019 MaidSafe.net limited.\n\
          //\n\
          // This SAFE Network Software is licensed to you under the MIT license\n\
          // <LICENSE-MIT or https://opensource.org/licenses/MIT> or the Modified\n\
