+++ safe_authenticator/src/apps.rs
@@ -11,16 +11,12 @@
 use super::{config, AuthFuture};
 use crate::app_auth::{app_state, AppState};
 use crate::client::AuthClient;
-use crate::ffi::apps as ffi;
-use crate::ffi::apps::AppPermissions as FfiAppPermissions;
-use crate::ffi::apps::RegisteredApp as FfiRegisteredApp;
 use crate::{app_container, AuthError};
 use bincode::deserialize;
-use ffi_utils::{vec_into_raw_parts, ReprC};
 use futures::future::Future;
 use safe_core::client::{AuthActions, Client};
 use safe_core::core_structs::{access_container_enc_key, AccessContainerEntry, AppAccess};
-use safe_core::ipc::req::{containers_from_repr_c, containers_into_vec, ContainerPermissions};
+use safe_core::ipc::req::ContainerPermissions;
 use safe_core::ipc::{AppExchangeInfo, IpcError};
 use safe_core::utils::symmetric_decrypt;
 use safe_core::FutureExt;
@@ -37,54 +33,6 @@ pub struct RegisteredApp {
     pub containers: HashMap<String, ContainerPermissions>,
     /// Permissions allowed for the app
     pub app_perms: AppPermissions,
-}
-
-impl RegisteredApp {
-    /// Construct FFI wrapper for the native Rust object, consuming self.
-    pub fn into_repr_c(self) -> Result<FfiRegisteredApp, IpcError> {
-        let Self {
-            app_info,
-            containers,
-            app_perms,
-        } = self;
-
-        let container_permissions_vec = containers_into_vec(containers.into_iter())?;
-
-        let (containers_ptr, containers_len) = vec_into_raw_parts(container_permissions_vec);
-
-        let ffi_app_perms = FfiAppPermissions {
-            transfer_coins: app_perms.transfer_coins,
-            get_balance: app_perms.get_balance,
-            perform_mutations: app_perms.perform_mutations,
-        };
-
-        Ok(FfiRegisteredApp {
-            app_info: app_info.into_repr_c()?,
-            containers: containers_ptr,
-            containers_len,
-            app_permissions: ffi_app_perms,
-        })
-    }
-}
-
-impl ReprC for RegisteredApp {
-    type C = *const ffi::RegisteredApp;
-    type Error = IpcError;
-
-    #[allow(unsafe_code)]
-    unsafe fn clone_from_repr_c(repr_c: Self::C) -> Result<Self, Self::Error> {
-        let native_app_perms = AppPermissions {
-            transfer_coins: (*repr_c).app_permissions.transfer_coins,
-            get_balance: (*repr_c).app_permissions.get_balance,
-            perform_mutations: (*repr_c).app_permissions.perform_mutations,
-        };
-
-        Ok(Self {
-            app_info: AppExchangeInfo::clone_from_repr_c(&(*repr_c).app_info)?,
-            containers: containers_from_repr_c((*repr_c).containers, (*repr_c).containers_len)?,
-            app_perms: native_app_perms,
-        })
-    }
 }
 
 /// Removes an application from the list of revoked apps.
