+++ safe_authenticator/src/lib.rs
@@ -32,14 +32,6 @@
 
 // Public exports. See https://github.com/maidsafe/safe_client_libs/wiki/Export-strategy.
 
-// Export FFI interface
-
-pub use crate::ffi::apps::*;
-pub use crate::ffi::errors::codes::*;
-pub use crate::ffi::ipc::*;
-pub use crate::ffi::logging::*;
-pub use crate::ffi::*;
-
 // Export public auth interface.
 
 pub use self::errors::AuthError;
@@ -51,7 +43,6 @@ pub mod app_container;
 pub mod apps;
 pub mod config;
 pub mod errors;
-pub mod ffi;
 pub mod ipc;
 pub mod revocation;
 #[cfg(any(test, feature = "testing"))]
@@ -62,7 +53,7 @@ mod std_dirs;
 #[cfg(test)]
 mod tests;
 
-use crate::ffi::errors::Error;
+use errors::{AuthError as Error, Result as AuthResult};
 use futures::stream::Stream;
 use futures::sync::mpsc;
 use futures::{Future, IntoFuture};
@@ -106,7 +97,7 @@ pub struct Authenticator {
 
 impl Authenticator {
     /// Send a message to the authenticator event loop.
-    pub fn send<F>(&self, f: F) -> crate::ffi::errors::Result<()>
+    pub fn send<F>(&self, f: F) -> AuthResult<()>
     where
         F: FnOnce(&AuthClient) -> Option<Box<dyn Future<Item = (), Error = ()>>> + Send + 'static,
     {
