+++ safe_authenticator_ffi/cargo.toml
@@ -1,11 +1,12 @@
 [package]
 authors = ["MaidSafe Developers <dev@maidsafe.net>"]
+build = "build.rs"
 edition = "2018"
 description = "SAFE Authenticator"
 documentation = "https://docs.rs/safe_authenticator"
 homepage = "https://maidsafe.net"
 license = "GPL-3.0"
-name = "safe_authenticator"
+name = "safe_authenticator_ffi"
 readme = "README.md"
 repository = "https://github.com/maidsafe/safe_client_libs"
 version = "0.13.0"
@@ -17,9 +18,11 @@ futures = "~0.1.17"
 jni = { version = "~0.12.0", optional = true }
 log = "~0.4.1"
 lru-cache = "~0.1.1"
+quic-p2p = "~0.3.0"
 rand = "0.6"
 safe_core = { path = "../safe_core", version = "~0.38.1" }
-safe-nd = "~0.8.0"
+safe_authenticator = { path = "../safe_authenticator", version = "~0.13.0" }
+safe-nd = "~0.6.2"
 serde = { version = "~1.0.97", features = ["derive"] }
 threshold_crypto = "~0.3.2"
 tiny-keccak = "~1.5.0"
@@ -30,12 +33,13 @@ env_logger = { version = "~0.6.2", optional = true }
 [dev-dependencies]
 env_logger = "~0.6.2"
 safe_core = { path = "../safe_core", version = "~0.38.1", features = ["testing"] }
+safe_authenticator = { path = "../safe_authenticator", version = "~0.13.0" }
 
 [build-dependencies]
 ffi_utils = "~0.15.0"
 jni = "~0.12.0"
 safe_bindgen = { version = "~0.13.2", optional = true }
-safe-nd = "~0.8.0"
+safe-nd = "~0.6.2"
 threshold_crypto = "~0.3.2"
 unwrap = "~1.2.0"
 
@@ -43,3 +47,6 @@ unwrap = "~1.2.0"
 mock-network = ["testing", "safe_core/mock-network"]
 testing = ["safe_core/testing", "env_logger"]
 bindings = ["safe_bindgen"]
+
+[lib]
+crate_type = ["staticlib", "rlib", "cdylib"]
