+++ safe_authenticator_ffi/src/ffi/errors/codes.rs
