+++ safe_authenticator_ffi/src/ffi/mod.rs
@@ -14,18 +14,19 @@
 pub mod apps;
 /// Errors
 pub mod errors;
+pub mod helpers;
 /// Authenticator communication with apps
 pub mod ipc;
 /// Logging utilities
 pub mod logging;
 
 use crate::ffi::errors::{Error, Result};
-use crate::Authenticator;
 use ffi_utils::try_cb;
 use ffi_utils::{catch_unwind_cb, FfiResult, OpaqueCtx, ReprC, FFI_RESULT_OK};
 use log::trace;
 use rand::thread_rng;
-use safe_core::{config_handler, test_create_balance, Client};
+use safe_authenticator::Authenticator;
+use safe_core::{config_handler, test_create_balance};
 use safe_nd::{ClientFullId, Coins};
 use std::ffi::{CStr, OsStr};
 use std::os::raw::{c_char, c_void};
@@ -72,7 +73,6 @@ pub unsafe extern "C" fn create_acc(
             FFI_RESULT_OK,
             Box::into_raw(Box::new(authenticator)),
         );
-
         Ok(())
     })
 }
@@ -122,17 +122,19 @@ pub unsafe extern "C" fn auth_reconnect(
     user_data: *mut c_void,
     o_cb: extern "C" fn(user_data: *mut c_void, result: *const FfiResult),
 ) {
-    catch_unwind_cb(user_data, o_cb, || -> Result<_> {
+    catch_unwind_cb(user_data, o_cb, || -> Result<()> {
         let user_data = OpaqueCtx(user_data);
         (*auth).send(move |client| {
-            try_cb!(
-                client.restart_network().map_err(Error::from),
-                user_data.0,
-                o_cb
-            );
+            // Todo: The function isn't public
+            // try_cb!(
+            //     client.restart_network().map_err(Error::from),
+            //     user_data.0,
+            //     o_cb
+            // );
             o_cb(user_data.0, FFI_RESULT_OK);
             None
-        })
+        });
+        Ok(())
     })
 }
 
@@ -170,8 +172,6 @@ pub extern "C" fn auth_is_mock() -> bool {
 mod tests {
     use super::*;
     use crate::ffi::auth_is_mock;
-    use crate::run;
-    use crate::AuthError;
     use ffi_utils::test_utils::call_1;
     use futures::Future;
     use safe_core::{client::COST_OF_PUT, utils, FutureExt};
