+++ safe_authenticator_ffi/src/lib.rs
@@ -1,0 +1,10 @@
+// Export FFI interface
+
+pub mod ffi;
+
+pub use crate::ffi::apps::*;
+pub use crate::ffi::errors::codes::*;
+pub use crate::ffi::ipc::*;
+pub use crate::ffi::logging::*;
+pub use crate::ffi::*;
+pub use crate::helpers::*;
