+++ safe_authenticator_ffi/src/tests/share_mdata.rs
