+++ tests/Cargo.toml
@@ -12,9 +12,10 @@ version = "0.1.0"
 ffi_utils = "~0.15.0"
 futures = "~0.1.17"
 safe_app = { path = "../safe_app", version = "~0.13.0", features = ["testing"] }
+safe_authenticator_ffi = { path = "../safe_authenticator_ffi", version = "~0.13.0", features = ["testing"] }
 safe_authenticator = { path = "../safe_authenticator", version = "~0.13.0", features = ["testing"] }
 safe_core = { path = "../safe_core", version = "~0.38.1", features = ["testing"] }
-safe-nd = "~0.8.0"
+safe-nd = "~0.6.2"
 serde = "~1.0.24"
 serde_derive = "~1.0.24"
 serde_json = "~1.0.2"
@@ -24,5 +25,6 @@ unwrap = "~1.2.0"
 mock-network = [
     "safe_core/mock-network",
     "safe_authenticator/mock-network",
+    "safe_authenticator_ffi/mock-network",
     "safe_app/mock-network",
 ]
